wip: split up test runs
