+++ GasProfile.json
@@ -1,8 +1,8 @@
 {
-  "reduceToNewNoRefund": 59764,
-  "reduceToExistingNoRefund": 44754,
-  "emptyToExistingNoRefund": 28973,
-  "emptyToNewNoRefund": 44838,
+  "reduceToNewNoRefund": 59771,
+  "reduceToExistingNoRefund": 44761,
+  "emptyToExistingNoRefund": 28980,
+  "emptyToNewNoRefund": 44845,
   "reduceApprovalReducingToNew": 68725,
   "reduceApprovalReducingToExisting": 53715,
   "reduceApprovalEmptyingToNew": 53799,
@@ -14,9 +14,9 @@
   "approve50": 47590,
   "reduceToBurn": 49515,
   "emptyToBurn": 34589,
-  "burnMicroDollar": 49605,
-  "reduceToBurnWithChange": 49669,
-  "emptyToBurnWithChange": 49669,
+  "burnMicroDollar": 49515,
+  "reduceToBurnWithChange": 49579,
+  "emptyToBurnWithChange": 49579,
   "reduceApprovalReducingToBurn": 58563,
   "emptyApprovalReducingToBurn": 43637,
   "reduceApprovalEmptyingToBurn": 43573,
