+++ tests/BurnableToken.test.js
