+++ tests/BurnableTokenWithBounds.test.js
