+++ tests/CompliantToken.test.js
