+++ tests/DelegateERC20.test.js
@@ -1,7 +1,7 @@
 const CanDelegate = artifacts.require('CanDelegateMock')
 const TrueUSD = artifacts.require('TrueUSDMock')
-import standardTokenTests from './token/StandardToken';
-import basicTokenTests from './token/BasicToken';
+import standardTokenTests from './StandardToken';
+import basicTokenTests from './BasicToken';
 import compliantTokenTests from './CompliantToken';
 import depositTokenTests from './DepositToken'
 import redeemTokenTests from './RedeemToken';
