+++ tests/ProxyWithController.test.js
