+++ tests/ProxyWithTUSD.test.js
@@ -1,7 +1,7 @@
 import assertRevert from './helpers/assertRevert'
 import expectThrow from './helpers/expectThrow'
-import basicTokenTests from './token/BasicToken';
-import standardTokenTests from './token/StandardToken';
+import basicTokenTests from './BasicToken';
+import standardTokenTests from './StandardToken';
 const Registry = artifacts.require("RegistryMock")
 const TrueUSD = artifacts.require("TrueUSDMock")
 const TusdProxy = artifacts.require("OwnedUpgradeabilityProxy")
