+++ tests/StandardToken.test.js
