+++ tests/TokenWithHooks.test.js
