+++ tests/TrueCurrencies.test.js
