+++ tests/UpgradeHelper.test.js
